+++ KD_Lib/noisy/train.py
@@ -1,7 +1,12 @@
 import torch
+import torch.nn as nn
+import torch.optim as optim
 import torch.nn.functional as F
-from KD_Lib.noisy.utils import add_noise
+
 import random
+
+from KD_Lib.noisy.utils import add_noise, eval
+from KD_Lib.RKD import RKDLoss
 
 
 def train_teacher(model, train_loader, optimizer, epochs):
@@ -61,3 +66,28 @@ def train_student(teacher_model, student_model, train_loader, optimizer,
 
         loss_arr.append(epoch_loss)
         print(f'Epoch {e+1} loss = {epoch_loss}')
+
+def run_experiment(train_loader, test_loader, teacher_model, student_model, 
+                   epochs, lr, optimizer, loss, alpha, variance,
+                   rkd_dist, rkd_angle):
+
+    if optimizer.upper() == 'SGD':
+        t_optimizer = optim.SGD(teacher_model.parameters(), lr, momentum=0.9)
+        s_optimizer = optim.SGD(student_model.parameters(), lr, momentum=0.9)
+    elif optimizer.upper() == 'Adam':
+        t_optimizer = optim.Adam(teacher_model.parameters(), lr)
+        s_optimizer = optim.Adam(student_model.parameters(), lr)
+
+    if loss.upper() == 'MSE':
+        loss_fn = nn.MSELoss()
+    elif loss.upper() == 'KL':
+        loss_fn = nn.KLDivLoss()
+    elif loss.upper() == 'RKD':
+        loss_fn = RKDLoss(dist_ratio=rkd_dist, angle_ratio=rkd_angle)
+
+    train_teacher(teacher_model, train_loader, t_optimizer, epochs)
+    eval(teacher_model, test_loader)
+
+    train_student(teacher_model, student_model, train_loader, s_optimizer,
+                  loss_fn, epochs, alpha, variance)
+    eval(student_model, test_loader)
